Move ModelingLanguageChanged event
It belongs to te ModelingLanguage service.
